Extract RangeSlider from ProbabilitySlider
Generalize ProbabilitySlider into a RangeSlider with
configurable min/max props. ProbabilitySlider becomes a
thin wrapper (min=1, max=100). Enables reuse for gain
locks (min=0, max=100) in upcoming StepPopover work.
